various bugfixes and /claim
fixed confirmationStates not updating when the confirmation didn't timeout

went back to using Date.now instead of ntp, as this will  be more reliable once the bot is on its own server

added code to download wiki images of characters without cards.

added a claim command.

notify the user when they have 1 roll left.

consume a claim when upgrading to a higher rarity.